+++ src/main/java/ru/school/retailanalitycs_web_java/exceptions/notFoundExceptions/CardNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.school.retailanalitycs_web_java.exceptions.notFoundExceptions;
 
 public class CardNotFoundException extends EntityNotFoundException {
-    public CardNotFoundException(Integer id) {
+    public CardNotFoundException(Long id) {
         super(String.format("Card with id {%s} not found", id));
 
     }
