+++ src/main/java/ru/school/retailanalitycs_web_java/exceptions/notFoundExceptions/CustomerNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.school.retailanalitycs_web_java.exceptions.notFoundExceptions;
 
 public class CustomerNotFoundException extends EntityNotFoundException {
-    public CustomerNotFoundException(int id) {
+    public CustomerNotFoundException(Long id) {
         super(String.format("Customer with id {%d} not found", id));
     }
 }
