+++ src/main/java/ru/school/retailanalitycs_web_java/exceptions/notFoundExceptions/CustomerViewNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.school.retailanalitycs_web_java.exceptions.notFoundExceptions;
 
 public class CustomerViewNotFoundException extends EntityNotFoundException {
-    public CustomerViewNotFoundException(int id) {
+    public CustomerViewNotFoundException(Long id) {
         super(String.format("Customer View with id {%d} not found", id));
     }
 }
