+++ src/main/java/ru/school/retailanalitycs_web_java/services/viewServices/CustomerViewService.java
@@ -5,7 +5,7 @@ import ru.school.retailanalitycs_web_java.entities.views.CustomerView;
 import ru.school.retailanalitycs_web_java.repositories.viewRepositories.CustomerViewRepository;
 
 @Service
-public class CustomerViewService extends ReadOnlyBaseService<CustomerView, Integer> {
+public class CustomerViewService extends ReadOnlyBaseService<CustomerView, Long> {
     protected CustomerViewService(CustomerViewRepository repository) {
         super(repository);
     }
