+++ src/main/resources/db/migrations/db.changelog-1.0-addTables.sql
@@ -3,7 +3,7 @@
 --changeset maslynem:1
 CREATE TABLE IF NOT EXISTS personal_data
 (
-    Customer_ID            SERIAL PRIMARY KEY,
+    Customer_ID            BIGSERIAL PRIMARY KEY,
     Customer_Name          TEXT,
     Customer_Surname       TEXT,
     Customer_Primary_Email TEXT,
@@ -16,33 +16,33 @@ CREATE TABLE IF NOT EXISTS personal_data
 
 CREATE TABLE IF NOT EXISTS cards
 (
-    Customer_Card_ID SERIAL PRIMARY KEY,
-    Customer_ID      INT NOT NULL REFERENCES personal_data (Customer_ID)
+    Customer_Card_ID BIGSERIAL PRIMARY KEY,
+    Customer_ID      BIGINT NOT NULL REFERENCES personal_data (Customer_ID)
 );
 
 CREATE INDEX ON cards (customer_id);
 
 CREATE TABLE IF NOT EXISTS transactions
 (
-    Transaction_ID       SERIAL PRIMARY KEY,
-    Customer_Card_ID     INT              NOT NULL REFERENCES cards (Customer_Card_ID),
+    Transaction_ID       BIGSERIAL PRIMARY KEY,
+    Customer_Card_ID     BIGINT              NOT NULL REFERENCES cards (Customer_Card_ID),
     Transaction_Summ     DOUBLE PRECISION NOT NULL,
     Transaction_DateTime timestamp WITHOUT TIME ZONE,
-    Transaction_Store_ID INT
+    Transaction_Store_ID BIGINT
 );
 
 CREATE TABLE groups_sku
 (
-    Group_ID   SERIAL PRIMARY KEY,
+    Group_ID   BIGSERIAL PRIMARY KEY,
     Group_Name TEXT UNIQUE,
     CONSTRAINT proper_name CHECK (Group_Name ~ '^[a-zA-Zа-яА-Я0-9\s\-\_\.,:;!@#$%^&*()+=?"''<>\/\\\[\]\{\}\|]{1,}$')
 );
 
 CREATE TABLE sku
 (
-    SKU_ID   SERIAL PRIMARY KEY,
+    SKU_ID   BIGSERIAL PRIMARY KEY,
     SKU_Name TEXT,
-    Group_ID INT NOT NULL REFERENCES groups_sku (Group_ID),
+    Group_ID BIGINT NOT NULL REFERENCES groups_sku (Group_ID),
     CONSTRAINT proper_name CHECK (SKU_Name ~ '^[a-zA-Zа-яА-Я0-9\s\-\_\.,:;!@#$%^&*()+=?"''<>\/\\\[\]\{\}\|]{1,}$')
 );
 
@@ -50,8 +50,8 @@ CREATE INDEX ON sku (Group_ID);
 
 CREATE TABLE stores
 (
-    Transaction_Store_ID INT,
-    SKU_ID               INT              NOT NULL REFERENCES sku (SKU_ID),
+    Transaction_Store_ID BIGINT,
+    SKU_ID               BIGINT              NOT NULL REFERENCES sku (SKU_ID),
     SKU_Purchase_Price   DOUBLE PRECISION NOT NULL,
     SKU_Retail_Price     DOUBLE PRECISION NOT NULL,
     PRIMARY KEY (Transaction_Store_ID, SKU_ID)
@@ -59,8 +59,8 @@ CREATE TABLE stores
 
 CREATE TABLE IF NOT EXISTS checks
 (
-    Transaction_ID INT              NOT NULL REFERENCES transactions (Transaction_ID),
-    SKU_ID         INT              NOT NULL REFERENCES sku (SKU_ID),
+    Transaction_ID BIGINT              NOT NULL REFERENCES transactions (Transaction_ID),
+    SKU_ID         BIGINT              NOT NULL REFERENCES sku (SKU_ID),
     SKU_Amount     DOUBLE PRECISION NOT NULL,
     SKU_Summ       DOUBLE PRECISION NOT NULL,
     SKU_Summ_Paid  DOUBLE PRECISION NOT NULL,
